Limit menu div width to 300px
The goal now is to make the div not take up the entire width of the page. The CSS width property is perfect for this.

You can use the id selector to target a specific element with an id attribute. An id selector is defined by placing the hash symbol # directly in front of the element's id value. For example, if an element has the id of cat then you would target that element like this:

#cat {
  width: 250px;
}
Use the #menu selector to give your element a width of 300px.