Replace #menu selector with .menu class
So far you have been using type and id selectors to style elements. However, it is more common to use a different selector to style your elements.

A class selector is defined by a name with a dot directly in front of it, like this:

.class-name {
  styles
}
Change the existing #menu selector into a class selector by replacing #menu with a class named .menu.